Prevents submits when locked

+++ process.php
@@ -5,49 +5,58 @@ header('Content-Type: application/json');
 $now = date('Y-m-d H:i:s');
 $day = date('Y-m-d');
 
+$locked = (file_get_contents('closed.lock') == date('Y-m-d'));
+
 $path = dirname(__FILE__).'/main.sqlite3';
 
-if(isset($_POST['user'])){
+if(!$locked){
 
-	$user = $_POST['user'];
+	if(isset($_POST['user'])){
 
-	try {
-		$db = new SQlite3($path);
-		$begun = false;
-		
-		if($_POST['action'] == 'add'){
+		$user = $_POST['user'];
 
-			$db->query('BEGIN;');
-			$begun = true;
+		try {
+			$db = new SQlite3($path);
+			$begun = false;
+			
+			if($_POST['action'] == 'add'){
 
-			$db->query(sprintf("DELETE FROM pedidos WHERE user='%s' AND reference_day='%s';", $user, $day));
+				$db->query('BEGIN;');
+				$begun = true;
 
-			$fields = '(user, flavour, quantity, ordered_at, reference_day)';
-			foreach ($_POST['preferences'] as $pref){
-				$values = sprintf("('%s', '%s', %d, '%s', '%s')", $user, $pref['pizza'], $pref['pieces'], $now, $day);
-				$result = $db->query(sprintf('INSERT INTO pedidos %s VALUES %s ;', $fields, $values));
-				if(! $result){
-					throw new Exception("Insert Error ".$values, 2);
+				$db->query(sprintf("DELETE FROM pedidos WHERE user='%s' AND reference_day='%s';", $user, $day));
+
+				$fields = '(user, flavour, quantity, ordered_at, reference_day)';
+				foreach ($_POST['preferences'] as $pref){
+					$values = sprintf("('%s', '%s', %d, '%s', '%s')", $user, $pref['pizza'], $pref['pieces'], $now, $day);
+					$result = $db->query(sprintf('INSERT INTO pedidos %s VALUES %s ;', $fields, $values));
+					if(! $result){
+						throw new Exception("Insert Error ".$values, 2);
+					}
 				}
+				$db->query('COMMIT;');
+				$message = 'Pedido realizado, '.$_POST['user'].'! Agora é só esperar um pouquinho e ser feliz!';
+
+			} elseif ($_POST['action'] == 'delete'){
+
+				$db->query(sprintf("DELETE FROM pedidos WHERE user='%s' AND reference_day='%s';", $user, $day));
+				$message = 'Tudo bem, meu jovem, o tio já deletou seus pedidos. Até a próxima!';
+
 			}
-			$db->query('COMMIT;');
-			$message = 'Pedido realizado, '.$_POST['user'].'! Agora é só esperar um pouquinho e ser feliz!';
-
-		} elseif ($_POST['action'] == 'delete'){
-
-			$db->query(sprintf("DELETE FROM pedidos WHERE user='%s' AND reference_day='%s';", $user, $day));
-			$message = 'Tudo bem, meu jovem, o tio já deletou seus pedidos. Até a próxima!';
-
+		} catch (Exception $e) {
+			if($begun){
+				$db->query('ROLLBACK;');
+			}
+			$message = 'Erro ao salvar pedido :/ ('.$e.')'; 
 		}
-	} catch (Exception $e) {
-		if($begun){
-			$db->query('ROLLBACK;');
-		}
-		$message = 'Erro ao salvar pedido :/ ('.$e.')'; 
+
+		respond_message($message);
 	}
 
-	respond_message($message);
+} else {
+	respond_message('Pedidos fechados...');
 }
+
 
 function respond($data){
 	echo json_encode($data);
